Code cleanup, remove duplicate code and use separate components where possible
